Dev: introduce sdk and refactor cctp to its own file
