+++ packages/arb-token-bridge-ui/src/hooks/CCTP/fetchCCTPLimits.ts
@@ -2,7 +2,7 @@ import { readContract } from '@wagmi/core'
 import { CCTPSupportedChainId } from '../../state/cctpState'
 import { FiatTokenProxyAbi } from '../../util/cctp/FiatTokenProxyAbi'
 import { TokenMinterAbi } from '../../util/cctp/TokenMinterAbi'
-import { getContracts } from './useCCTP'
+import { getContracts } from '@/token-bridge-sdk/cctp'
 
 /**
  *
