+++ packages/arb-token-bridge-ui/src/util/TokenApprovalUtils.ts
@@ -2,7 +2,7 @@ import { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__fact
 import { MaxUint256 } from '@ethersproject/constants'
 import { Provider } from '@ethersproject/providers'
 import { BigNumber, constants, Signer } from 'ethers'
-import { getContracts } from '../hooks/CCTP/useCCTP'
+import { getContracts } from '@/token-bridge-sdk/cctp'
 import { CCTPSupportedChainId } from '../state/cctpState'
 import { fetchErc20L1GatewayAddress } from './TokenUtils'
